requirements.txt
streamlit
opencv-python-headless
numpy
joblib
scikit-image
scikit-learn
Pillow
